+++ internal/online/postgres/online_storage.go
@@ -1,4 +1,4 @@
-package main
+package postgres
 
 import (
 	"context"
