+++ internal/online/postgres/user_online_pair.go
@@ -1,4 +1,4 @@
-package main
+package postgres
 
 import (
 	"time"
